items + languages +quests 엔티티에 중복 방지를 위해 unique 삽입 login 페이지에 스타일 분리가 되어있지 않아 분리 login 리다이렉트 로직 변경 -> memberController + surveyController 수정 ex ) 설문조사 페이지에서 로그인이 필요해서 로그인 페이지로 보내졌을 경우, 로그인 성공 후 설문조사 페이지로 이동

+++ src/main/java/com/abc/doge/repository/LanguageRepository.java
@@ -3,6 +3,6 @@ package com.abc.doge.repository;
 import com.abc.doge.entity.Languages;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LanguageRepository extends JpaRepository<Languages, Integer> {
+public interface LanguageRepository extends JpaRepository<Languages, Long> {
 }
